remove codes from Restauarant component to Order component
